More standardizing
Renamed the vector classes from VectorN to VecN, and added more methods.

Also changed scalar keyframes from Double to Float.

+++ modelstudio/src/com/hiveworkshop/rms/editor/render3d/RenderParticleEmitter2View.java
@@ -19,7 +19,7 @@ public class RenderParticleEmitter2View extends EmitterView {
 		final AnimFlag emissionRateFlag = emitter.getEmitter().find("EmissionRate");
 		if (emissionRateFlag != null) {
 			if (emissionRateFlag.size() > 0) {
-				lastEmissionRate = ((Double) emissionRateFlag.getValues().get(0)).doubleValue();
+				lastEmissionRate = ((Float) emissionRateFlag.getValues().get(0)).floatValue();
 			}
 		}
 	}
